[MODI-79] REFACTOR: dto 패키지 구조 세분화
