+++ src/main/java/com/prgrms/modi/party/PartyService.java
@@ -4,8 +4,8 @@ import com.prgrms.modi.ott.domain.OTT;
 import com.prgrms.modi.ott.service.OttService;
 import com.prgrms.modi.party.domain.Party;
 import com.prgrms.modi.party.domain.PartyStatus;
-import com.prgrms.modi.party.dto.PartyListResponse;
-import com.prgrms.modi.party.dto.PartyResponse;
+import com.prgrms.modi.party.dto.response.PartyListResponse;
+import com.prgrms.modi.party.dto.response.PartyResponse;
 import java.time.LocalDate;
 import java.util.stream.Collectors;
 import org.springframework.data.domain.PageRequest;
